requirements.txt
pymongo
websockets
python-telegram-bot
PyYAML
jinja2
matplotlib
requests
